+++ UI/src/com/kotcrab/vis/ui/widget/VisCheckBox.java
@@ -104,9 +104,9 @@ public class VisCheckBox extends TextButton implements Focusable, BorderOwner {
 		super.draw(batch, parentAlpha);
 
 		if (stateInvalid && style.errorBorder != null)
-			style.errorBorder.draw(batch, getX(), getY() + image.getY(), image.getWidth(), image.getHeight());
+			style.errorBorder.draw(batch, getX() + image.getX(), getY() + image.getY(), image.getWidth(), image.getHeight());
 		else if (focusBorderEnabled && drawBorder && style.focusBorder != null)
-			style.focusBorder.draw(batch, getX(), getY() + image.getY(), image.getWidth(), image.getHeight());
+			style.focusBorder.draw(batch, getX() + image.getX(), getY() + image.getY(), image.getWidth(), image.getHeight());
 	}
 
 	public Image getImage () {