+++ UI/src/com/kotcrab/vis/ui/VisUI.java
@@ -31,7 +31,7 @@ import com.badlogic.gdx.utils.GdxRuntimeException;
  */
 public class VisUI {
 	/** Current VisUI version, does not include SNAPSHOT even if this version is snapshot. */
-	public static final String VERSION = "1.0.1";
+	public static final String VERSION = "1.0.2";
 
 	private static final String TARGET_GDX_VERSION = "1.9.2";
 	private static boolean setSkipGdxVersionCheck = false;