+++ UI/CHANGES.md
@@ -1,4 +1,5 @@
 #### Version: 1.0.1 (LibGDX 1.9.2)
+- **Fixed**: Not being able to resize window with `TabbedPane`
 
 #### Version: 1.0.0 (LibGDX 1.9.2)
 - **Changed**: `InputValidator` moved to `com.kotcrab.vis.ui.util` package