+++ UI/CHANGES.md
@@ -1,3 +1,5 @@
+#### Version: 1.0.2 (LibGDX 1.9.2)
+
 #### Version: 1.0.1 (LibGDX 1.9.2)
 - **Added**: `ListView#getListAdapter()`
 - **Added**: `ListView#rebuildView()` and `UpdatePolicy.MANUAL`